requirements.txt
Pillow
torch
torchvision
numpy
pandas
opencv-python
matplotlib